backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pandas==2.3.2
openpyxl==3.1.5
google-generativeai==0.8.5
python-dotenv==1.1.1
deep-translator==1.11.4
